Create stepper motor in Python
Write a Python program to run a stepper motor in Python with following conditions. Assume unipolar stepper motor. The step sequence should be '1010', '0110', '0101', '1001'. The term ‘1’ represents excitation of stepper motor winding. The motor has to move ‘forward’ and ‘reverse’ based on user request. Assume number of steps as 2. Print the winding sequence as well as direction i.e. forward of reverse.